requirements.txt
# Core
Django==5.0.9
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9

# Multi-tenancy
django-tenants==3.7.0

# REST Framework & JWT
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# CORS (for frontend integration)
django-cors-headers==4.4.0

# Production server
gunicorn==23.0.0
whitenoise==6.11.0  # Serve static files efficiently

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
